requirements.txt
numpy==1.26.4
pgzero==1.2.1
pillow==10.2.0
pygame==2.5.2
